pyproject.toml
[project]
name = "ragproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "langchain>=1.2.17",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.2",
    "langchain-mistralai>=1.1.3",
    "langchain-openai>=1.2.1",
    "lxml>=6.1.0",
    "mistralai>=2.4.4",
    "numpy>=2.4.4",
    "openai>=2.33.0",
    "pandas>=3.0.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.4.1",
    "streamlit>=1.57.0",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "unstructured>=0.22.26",
    "uvicorn>=0.46.0",
]
